now decodes integers and enums
Had a few minutes to kill.
I decided on a large integer representation here, so for integers I'll
just pass the raw bytes up to the next layer.
For enums, the representation of choice is int, so that's what I went
with.
Of course, if someone decides to pass key as integers (which they
quite often would) they'll have to know how to handle those bytes
correctly (sign extension being the most likely problem they might run
into).