parse bit strings
Because this is a DER decoder, which means sizes must always be
determinate, we only parse primitive bit strings.
We essentially ignore the constructed flag, though, as it's redundant
with the indeterminate length on constructed bit strings and I see no
way for an attacker to exploit setting the flash but not using an
indeterminate length and some encoders might (not unreasonably, but
wrongly nonetheless) always set the flash on bit strings.

Anecdotally, this code was written yesterday, in the car, while my
wife was driving :slightly_smiling_face: