first sketch of an encoder
The encoder looks like it'll be a collection of functions that output
the to-encode values as DER to an output iterator. An octet string is
easy enough to encode using this scheme: it's just matter of encoding
the type, then the length, them copying everything into the output.
That does mean, though, that I need more than a single-pass input
iterator (because I need to get the distance between the two iterators
to get there length).
I might also add a length-and-iterator or length-and-pointer because
for non-random-access iterators, getting the distance may be expensive.