continue a bit on decoding reals
It's a nice night to be coding on one's phone.

Special real values are now decoded, and the entire value is copied
into the buffer. Now we need to get the number of octets for the
exponent (which we'll need to check against the octets we actually
have) after which we can finish calculating the mantissa.

I've decided on using double to represent reals. I could make that
configurable, but I don't see much of a reason for doing that.

The new Integer class is now used to report the integer value.