Add documentation
The ASN.1 schema language, which defines the enclosed structure in
human-readable format.

Note that it's very unlikely that I'll implement all of these: there
is very little use for me to define classes or constraints in ASN.1
and there exists very little in terms of real-world use of these
particular parts of ASN.1. Moreover, almost everything useful
specified in X.68{1,2,3} can be expressed in X.680