At a few simple types to the encoder
Adding simple types, such as booleans and end-of-contents, is very easy
with this approach: as long as the output iterator is updated by
anything that writes to the output we simply push whatever we need
into the output by writing to it.

Note that an output iterator should be smart enough to know when it
can't write anymore: as back_inserter will do nicely; a pointer into a
fixed-size array is dangerous.

Caveat emptor, as the Romans used to say...