Encode enumerated values
Next step will be to encode bit strings. We'll need to decide on a
representation of those first, though, as the decoder doesn't have a
representation other than a pair of iterators and a bits-to-ignore
count at the moment.
Of course, were don't actually need anything more than a pair of
iterators and a bits-to-ignore count...