encode bit strings
Representing a bit string as a pair of iterators and an integer should
be sufficient in almost any case, and add we require (because this is
a DER encoder) for the entire bit string to be available, this'll do.

Note, again, that the expected input iterator is a multi-pass input
iterator: we need to be able to calculate the distance between the
beginning and the end of the input range.