encode reals
Real encoding is among the more complex encodings in DER. It doesn't
follow IEEE-754 at all, so we have to extract the mantissa and the
exponent, in the case of the mantissa decide whether to represent it
in base 2, 8 or 16 and in either case encode the sign and the value.
Subnormal values resolve to zero in the mantissa and can't
represented, so they will hit am assertion on line 175 of this code.

I should note that I heartily dislike floating points for various
reasons, not least of which quirks like subnormal values, epsilon,
whacky encodings, etc. Add to that that most, if not all, floating
point values "in the wild" are really either a function of some
irrational number (pi, e, ...), a fraction, or an integer measurement
with scaling (gain and offset) lossy representations like IEEE-754 or
this one are really unnecessary (just store the function, fraction, or
measurements in stead).