Some proof-of-concept code for reals
Decoding reals needs some refactoring: while I've been reading up on
IEEE-754 representation and the standard C++ functions for floating
point types (and confirming that you should avoid touching this stuff if
you can) I've come to the conclusion that writing this part of the code
on a tablet or phone is probably not going to lead anywhere, so I
decided to pull out my laptop and get some code running.

The Details::Integer class compiles and runs - at least as far as this
proof of concept is concerned. The POC consists of a brand new
implementation of dissecting and building floating point values (from or
into double, resp.). I've tested it against an independent
implementation (the calculator on my phone) and arrived at the
conclusion that at least for the two values I've tried so far, it is
correct. That leads me to believe the method itself is correct as well.

In this commit, we also add the dependencies we'll be working with from
here on out: CMake for building (and my cmake submodule for the shared
configuration stuff) and the exceptions library with its support for my
contract-theory macros (pre-conditions, invariants, etc.).

I will use the two new functions in the encoder and decoder, replacing
part of what's there now.