Encode and decode sequences and sets
Decoding sequences and sets is pretty much a question of trying to keep
track of where you are in the encoded sequence or set (the only
difference between the two being that the order in a sequence is
significant whereas it isn't in a set). Once you run out of input octets
for the sequence or set, you're done with it. At this level, we don't
need to know what was expected in terms of the contents of the sequence
or set.
Hence, we have a simple pair of stacks: one for the type (sequence or
set) and one for the length of the sequence or set being decoded. We
allow the stack size to be configured, but the default (16) seems
reasonable to me. Whenever we decode a sequence or set, we simply push
the remaining octets for that sequence or set, and its type, onto the
stacks and wait for them to run out. Because DER doesn't support
indeterminate lengths, that's all we need to do. If the length of the
sequence and the length of the contents don't match up, that will
inevitably result in a decoding error. That also addresses the
"syntactic poison" Adam (@jadamcrain) pointed out in
https://twitter.com/jadamcrain/status/891984403706130432

For the encoder, things are slightly more painful: because DER doesn't
allow for indeterminate lengths, we have to know the length of whatever
it is we're encoding, which in the case of a sequence depends on the
contents of that sequence. The only way I can deal with that at the
moment is to pretend whoever is calling the encoder will encode into a
vector or buffer of some sort, and be able to pass us the iterators into
that vector or buffer so we can insert it into whatever it is we're
encoding into. This makes sense w.r.t. the way most of these things are
generally structured, but it does impose a serious limitation on whoever
would end up using this code.

Now for some code generation.