Rudimentary preprocessor
Mostly to show how a lexer works.

A lexer, or lexicographical scanner, reads the input and transforms it
into tokens. In our case, those tokens are begin and end comment ('/*'
and '*/'), Ada-style comment delimiters ('--'), newlines, quotes and
characters.
We need the quotes to be able to ignore comments within quotes (though I
haven't read X.680 in detail enough yet to know whether that is really
necessary).

Note that this little preprocessor is extremely inefficient: every token
text is a string, so we have to construct a string for every single
character we read. That doesn't matter much at this point, but we will
almost certainly forego using this preprocessor further along, when we
have a more interesting way to ignore comments by just parsing them,
then snipping them out of the parse tree.
The inefficiency is only due to the way Antlr generalizes tokens (I
would personally have handed back a pair of iterators into a parse
buffer, rather than a string, for every token).