Started work on the compiler
I've renamed the preprocessor to asn1p to be more consistent with what
the compiler will be called (asn1c) and started work on the compiler by
creating a listener to gather information from the parse tree.

I've also added a dependency for my little tracing library (which I've
put on GitHub for this project, and made open source) as that will help
with outputting traces etc.