Picking up where I left off
As things go, it's been about six months since I last touched this
project. I was discussing a different project with a junior developer
this week that involved ASN.1-encoded objects and some cryptography (the
two go hand in hand a lot of the time) and decided to pick this project
back up.

A few things need to be parsed before we can get to code generation, and
this is the more boring part of the code, so I might as well try to get
it over with...