Resolve defined types
We need to resolve these before we resolve anything else, and we need to
resolve all of them because resolution may pick them up from basically
anywhere in the tree afterwards, so it gets more difficult to handle if
we get there later. Hence, we add a whole phase to resolution and send a
new visitor throught the tree to do the resolution for us.

While testing, it turned out these defined types can be a bit sneaky in
terms of where they hide in the tree, so I may still have missed some,
thought I think I caught all of them.

All the visitor does is annotate the defined type with the type it
references. We don't handle imported types yet, because we can't export
types yet (I'll need to figure out a format to export the encoding state
machine when I get there, which I'll then need to import when I import
the types. That oughta be fun!)